Bug fix edit delete ownership issue

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,25 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        // Error message in case of non-owner trying to edit the image
+        String error = "Only the owner of the image can edit the image";
+        model.addAttribute("editError", error);
+
+        User user = imageService.getImage(imageId).getUser();
+        User sessionUser = (User) session.getAttribute("loggeduser");
+        if(user!= null && user.getId().equals(sessionUser.getId())) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else {
+            List<Tag> tags = image.getTags();
+            model.addAttribute("tags", tags);
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +144,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/";
     }
 
 
@@ -140,9 +152,23 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        List<Tag> tags = image.getTags();
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+
+        String error = "Only the owner of the image can delete the image";
+        model.addAttribute("deleteError", error);
+
+        User user = imageService.getImage(imageId).getUser();
+        User sessionUser = (User) session.getAttribute("loggeduser");
+        if(user.getId().equals(sessionUser.getId())) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            return "images/image";
+        }
     }
